refactor(app): extract cross-cutting app-shell infrastructure to core_app_shell
Phase 2 continuation of the SSOT migration (ADR-0011): move 4 modules
with no app-local UI dependencies into a new core_app_shell package —
app theme provider, deferred startup-task scheduling, platform-feature
configuration, and device form-factor detection.

app/lib's old paths become one-line re-export shims so relative-import
consumers within app/lib need no changes. 5 test files using absolute
package:airo_app/... imports were updated to package:core_app_shell/...
directly.

global_error_handler.dart was evaluated but excluded from this pass:
it depends on app/lib/shared/widgets/bug_report_dialog.dart and
app/lib/core/utils/logger.dart, both still app-local UI/utility code,
so extracting it now would either drag those along (out of scope) or
force core_app_shell to depend on app (violates its module.yaml
contract). Left in place for a future pass once those dependencies are
themselves addressed.

Verified: core_app_shell analyzes clean (info-level style lints only);
app analyze/tests for all touched files pass; the one failing test
(main_tv_debug_playlist_test.dart) fails identically on main due to a
pre-existing, unrelated gap — firebase_options.dart is gitignored and
not present in this worktree, so main_tv.dart fails to compile
regardless of this change (confirmed via git stash against baseline).
scripts/check-module-manifests.py: 57/57 valid.
