+++ packages/core_app_shell/module.yaml
@@ -1,0 +1,10 @@
+name: core_app_shell
+owner: Chief Architect
+reviewers:
+  - Flutter Architect
+  - Chief QA Officer
+allowed_dependencies:
+  - core_ui
+forbidden_dependencies:
+  - app
+quality_gates: {}
